requirements.txt
django==4.1.3
djangorestframework==3.14.0
XlsxWriter==3.0.3
environs==9.5.0
dj-config-url==0.1.1
